AccessLogUpdate

+++ konekcija/Form1.cs
@@ -14,7 +14,8 @@ namespace konekcija
         //lokacija Access baze 
         //ova lokacija se kooristi u fji GetTableNames() kao argumetn OleDbConnection() fje 
         //ovdje je trebalo instalirati microsoft Access Database Engine 2010 ali za 32bitni sistem tacnije sa ovog linka :https://www.microsoft.com/en-us/download/confirmation.aspx?id=13255
-        const string connectionString = "Provider=Microsoft.ACE.OLEDB.12.0; Data Source=C:\\ASManager\\ASRes\\ASConfig.mdb;Persist Security Info = False;";
+        const string connectionStringConf = "Provider=Microsoft.ACE.OLEDB.12.0; Data Source=C:\\ASManager\\ASRes\\ASConfig.mdb;Persist Security Info = False;";
+        const string connectionStringLog = "Provider=Microsoft.ACE.OLEDB.12.0; Data Source=C:\\ASManager\\ASRes\\ASLog.mdb;Persist Security Info = False;";
 
         // lista(spisak) tabela u Access bazi, ova lista je nadalje vidljiva i u funkciji koja sakuplja imena tabela GetTableNames();
         List<string> tables = new List<string>();
@@ -28,7 +29,6 @@ namespace konekcija
             //dodavanje Event handlera na klik dugmeta button1 
             Button1.Click += new EventHandler(Button1_Click);
         }
-
         //akcije na klik dugmeta
         private void Button1_Click(object sender, EventArgs e)
         {
@@ -37,19 +37,17 @@ namespace konekcija
             labela.Text = "Hello ";
             labela.Refresh();
             MessageBox.Show("zdravo");
-
             //sakupljanje imena tabela
-            GetTableNames();
-
+            tables.Clear();
+            GetTableNames(connectionStringConf);
             //komunikacija izmedju dvije baze
             //lokacija za SQL server 
             const string connectionStringDest = "Data Source = PD; Initial Catalog=Moja;Integrated Security=SSPI; MultipleActiveResultSets=true;";
 
-            using (sourceConnection = new OleDbConnection(connectionString))
+            using (sourceConnection = new OleDbConnection(connectionStringConf))
             {
                 //otvaranje izvorne baze Access
                 sourceConnection.Open();
-
                 //konektovanje na SQL bazu
                 using (destinationConnection = new SqlConnection(connectionStringDest))
                 {
@@ -59,52 +57,146 @@ namespace konekcija
                     MessageBox.Show("SQL baza otvorena");
                     foreach (string tbl in tables)
                     {
-                        if (tbl == "Cardholder")
+                        switch (tbl)
                         {
-                            Cardholder_update(sourceConnection, destinationConnection, tbl);
-                            return;
-                        }
-                        else
-                        {
-                            if (tbl == "AccessLog")
-                            {
-                                Logs_update(sourceConnection, destinationConnection, tbl);
-                                return;
-                            }
-                            else
-                            {
-                                MessageBox.Show("nije cardholder ni AccessLog");
-                            }
+                            case "Cardholder":
+                                Cardholder_update(sourceConnection, destinationConnection, tbl);
+                                break;
+                            default:
+                                Console.WriteLine("Default case");
+                                break;
                         }
                     }
                     sourceConnection.Close();
                     destinationConnection.Close();
                 }
             }
+
+            tables.Clear();
+            GetTableNames(connectionStringLog);
+            
+            using (sourceConnection = new OleDbConnection(connectionStringLog))
+            {
+                //otvaranje izvorne baze Access
+                sourceConnection.Open();
+                //konektovanje na SQL bazu
+                using (destinationConnection = new SqlConnection(connectionStringDest))
+                {
+                    //otvaranje SQL baze
+                    destinationConnection.Open();
+
+                    MessageBox.Show("SQL baza otvorena");
+                    foreach (string tbl in tables)
+                    {
+                        switch (tbl)
+                        {
+                            case "AccessLog":
+                                Logs_update(sourceConnection, destinationConnection, tbl);
+                                MessageBox.Show("AccessLog");
+                                break;
+                            default:
+                                Console.WriteLine("Default case");
+                                break;
+                        }
+                    }
+                    sourceConnection.Close();
+                    destinationConnection.Close();
+                }
+            }
+            
         }
 
-            public List<string> GetTableNames()
+            public List<string> GetTableNames(string connectionString)
             {
                 try
                 {
                     using (OleDbConnection con = new OleDbConnection(connectionString))
                     {
                         con.Open();
-
-                        //shema koja sadrzi imena tabela u Access bazi
+                        //shema koja sadrzi imena tabela u Access bazi , 
                         foreach (DataRow r in con.GetSchema("Tables").Select("TABLE_TYPE = 'TABLE'"))
                         {
                             tables.Add(r["TABLE_NAME"].ToString());
                         }
                         return tables;
-                    }
+                        con.Close();
                 }
+                    
+            }
                 catch (Exception ex) { MessageBox.Show(ex.Message); }
                 return tables;
             }
 
-            public void Logs_update(OleDbConnection sourceConnection, SqlConnection destinationConnection, string tbl) {
-                return;
+            public void Logs_update(OleDbConnection sourceConnectionLog, SqlConnection destinationConnection, string tbl) {
+
+            MessageBox.Show("Logs ulaz");
+            // m1 je maximalni CardholderID a n1 je broj redova u Access tabeli Cardholder
+            var commandSourceData1 = new OleDbCommand("Select MAX(LogID) from " + tbl, sourceConnectionLog);
+            var commandSourceData2 = new OleDbCommand("Select COUNT(LogID) from " + tbl, sourceConnectionLog);
+            int m1 = (int)commandSourceData1.ExecuteScalar();
+            int n1 = (int)commandSourceData2.ExecuteScalar();
+            MessageBox.Show("maksimalni " + m1 + " ,broj redova " + n1);
+
+            // m1 je maximalni CardholderID a n1 je broj redova u Access tabeli Cardholder
+            var commandDestinationData1 = new SqlCommand("Select MAX(LogID) from " + tbl, destinationConnection);
+            var commandDestinationData2 = new SqlCommand("Select COUNT(LogID) from " + tbl, destinationConnection);
+            int n2 = (int)commandDestinationData2.ExecuteScalar();
+
+            //kopiranje
+            using (var bulkCopy = new SqlBulkCopy(destinationConnection))
+            {
+                //ako je tabela u SQL-u NIJE prazna
+                //problem -----> ako je prazna ne moze da se nadje maximalni CardholderID 
+                if (n2 != 0)
+                {
+                    //ako je prazna ne moze da se nadje maximalni CardholderID , pa se tek sada racuna
+                    int m2 = (int)commandDestinationData1.ExecuteScalar();
+                    MessageBox.Show(" maksimalni " + m2 + " ,broj redova " + n2);
+
+                    // ako broj redova NIJE isti i ako je m1 u accessu veci od m2u SQL kopiraju se oni koji su novi   
+                    if (n1 != n2 && m1 > m2)
+                    {
+                        MessageBox.Show("ima novih kardholdera,kopira se");
+                        // iz  access tablel izvlace se podaci(korisnici) koji su novi 
+                        var commandSourceData = new OleDbCommand("Select LogID, CardType, CardholderID, ControlerID, Direction, LocalTime from " + tbl + " where LogID <=" + @m1 + " and LogID >" + @m2, sourceConnectionLog);
+
+                        // ispisuje se CardholderID koji se dodaje u tabelu
+                        var commandSourceData3 = new OleDbCommand("Select LogID from " + tbl + " where LogID <=" + @m1 + "and LogID >" + @m2, sourceConnectionLog);
+                        int c = (int)commandSourceData3.ExecuteScalar();
+                        MessageBox.Show(" Ubacje se LogID " + c);
+
+                        //cita se iz izvorne Access tabele
+                        var reader = commandSourceData3.ExecuteReader();
+                        //ime tabele u SQL bazi
+                        bulkCopy.DestinationTableName = "dbo." + tbl;
+                        //pisanje u tabelu 
+                        try {
+                            MessageBox.Show("");
+                            bulkCopy.WriteToServer(reader); }
+                        catch (Exception ex) { MessageBox.Show(ex.Message); }
+                        finally { reader.Close(); }
+                    }
+                    //brisanje ovdje nije potrebno. ali treba razmotriti slucaj kada vise ne psotji cardholderID,tj kada je izbrisan, zbog relacija
+                    // ili podestiri automacko brisanje ili modifikovanje tabele prilikom brisanja Cardholdera
+                    else
+                    { }
+                }
+                else
+                {
+                    //ako je tabela SQL prazna sve kopiramo 
+                    var commandSourceData = new OleDbCommand("Select LogID, CardCode, CardType, CardholderID, ControllerID, Direction, LocalTime from " + tbl, sourceConnectionLog);
+                    var reader = commandSourceData.ExecuteReader();
+                    MessageBox.Show("tabela prazna ----> ima novih logova,kopira se");
+                    bulkCopy.DestinationTableName = "dbo." + tbl;
+                    try { bulkCopy.WriteToServer(reader); }
+                    catch (Exception ex) { MessageBox.Show(ex.Message); }
+                    finally { reader.Close(); }
+                }
+            }
+
+            MessageBox.Show("Logs izlaz");
+            return;
+
             }
 
             public void Cardholder_update(OleDbConnection sourceConnection, SqlConnection destinationConnection, string tbl) {
@@ -136,7 +228,6 @@ namespace konekcija
                         // ako broj redova NIJE isti i ako je m1 u accessu veci od m2u SQL kopiraju se oni koji su novi   
                         if (n1 != n2 && m1 > m2)
                         {
-
                             MessageBox.Show("ima novih kardholdera,kopira se");
                             // iz  access tablel izvlace se podaci(korisnici) koji su novi 
                             var commandSourceData = new OleDbCommand("Select CardholderID, Name, FirstName, LastName, Gender from " + tbl + " where CardholderID <=" + @m1 + " and CardholderID >" + @m2, sourceConnection);
@@ -147,14 +238,13 @@ namespace konekcija
                             MessageBox.Show(" Ubacje se CardholderID " + c);
 
                             //cita se iz izvorne Access tabele
-                            var reader = commandSourceData.ExecuteReader();
+                            var reader = commandSourceData3.ExecuteReader();
                             //ime tabele u SQL bazi
                             bulkCopy.DestinationTableName = "dbo." + tbl;
                             //pisanje u tabelu 
                             try { bulkCopy.WriteToServer(reader); }
                             catch (Exception ex) { MessageBox.Show(ex.Message); }
                             finally { reader.Close(); }
-
                         }
                         //brisanje
                         else
@@ -179,8 +269,7 @@ namespace konekcija
                     MessageBox.Show(" maksimalni " + m2 + " ,broj redova " + n2);
                     if (n1 < n2)
                     {
-                        // lista cardholderID-a iz SQL baze
-
+                        // lista  je lista cardholderID-a iz SQL baze
                         SqlDataReader rdr = null;
                         List<int> lista = new List<int>();
                         var commandDestinationData3 = new SqlCommand("Select CardholderID from " + tbl, destinationConnection);
@@ -190,8 +279,7 @@ namespace konekcija
                             // MessageBox.Show(" "+rdr[0]);
                             lista.Add(rdr.GetInt32(0));
                         }
-
-                        // lista  CardholderId iz Access baze
+                        // sourceLista je lista  CardholderId iz Access baze
                         OleDbDataReader Sourcerdr = null;
                         List<int> sourceLista = new List<int>();
                         var commandSourceData3 = new OleDbCommand("Select CardholderID from " + tbl, sourceConnection);
@@ -203,6 +291,8 @@ namespace konekcija
                         }
                         //for (int i = 0; i < sourceLista.Count; i++)
                         //MessageBox.Show(" " + sourceLista[i]);
+
+                        //svi oni koji su izbrisani iz access baze, tj oni kojih ima u sql ali ne u access,oni se posle brisu, izuzetak je CardholderID=-1 jer on je za neregistrovane korisnike kartica
                         var firstNotSecond = lista.Except(sourceLista).ToList();
                         MessageBox.Show(" proslo");
                         //var commandDestinationData4= new SqlCommand("Select * from "+tbl, destinationConnection);
@@ -212,6 +302,8 @@ namespace konekcija
                         SqlDataReader rdr1 = null;
                         for (int i = 0; i < firstNotSecond.Count; i++)
                         {
+                        if (firstNotSecond[i] != -1)
+                        {
                             MessageBox.Show(" " + firstNotSecond[i]);
                             int n = (int)firstNotSecond[i];
                             var command = new SqlCommand("Delete from " + tbl + " where CardholderID=" + @firstNotSecond[i], destinationConnection);
@@ -220,12 +312,14 @@ namespace konekcija
                                 rdr1 = command.ExecuteReader();
                                 MessageBox.Show("successfully data Deleted", "user information");
                                 while (rdr1.Read())
-                                {}
+                                { }
                             }
                             catch (Exception ec)
                             {
                                 MessageBox.Show(ec.Message);
                             }
+                        }
+                        else { }
                         }
                     }
                     else { }
